requirements.txt
fastapi
uvicorn
pathway
google-generativeai
python-dotenv
requests
